multi-GPU tests: capture trace on interrupt (#793)
## Describe your changes

`./run_test_with_accelerate.py` tees the output of pytest. But if you `ctrl-C` pytest when it's being teed, you don't see the pytest errors / trace, I think because the `tee` closes from the SIGINT before `pytest` runs its SIGINT handler that prints a stack trace. This is annoying when debugging. I can re-pro this by running `pytest --full-trace 2>&1 | tee /tmp/out.txt` and noticing that nothing is printed when I `ctrl-C` it, but if I remove the `2>&1 | tee ...` then I see a trace.

This PR adds the `-i` flag to `tee` that makes it ignore SIGINT, so instead `pytest` handles the SIGINT and closes its output upon dying, which is then what causes tee to close. (I wonder if there are other signals that I should ignore, but tee only has a flag for ignoring SIGINT. We could `trap` other signals if needed. I did check that `pytest` doesn't print a stack trace when it gets a SIGTERM even without tee, so that's not a signal we need to specially handle.)



## Issue ticket number and link

none

## Describe tests performed


## Checklist before requesting a review
- [x] I have performed a self-review of my code
- [ ] If it is a new feature, I have added thorough tests.
- [ ] I have considered how it would affect the multi-GPU case and tested on a multi-GPU node if appropriate.
- [ ] If it is a high-risk change, I have run the following experiments to verify that the results are as expected:
    - [ ] [experiment=ian/027_eval_refactored_random_token_imdb model.name_or_path=AlignmentResearch/robust_llm_pythia-imdb-1.4b-mz-ada-v3](https://wandb.ai/farai/robust-llm/runs/zlynwimq)
    - [ ] [experiment=ian/028_eval_refactored_gcg_imdb model.name_or_path=AlignmentResearch/robust_llm_pythia-imdb-2.8b-mz-ada-v3](https://wandb.ai/farai/robust-llm/runs/pi6eq0wp)
    - [ ] [experiment=niki/1_DEBUG_niki_128](https://wandb.ai/farai/robust-llm/runs/76h9beb3)