Trainer: assert number of GPUs per process is correct (#885)
## Describe your changes

We conflate number of devices & number of processes — for example, in `training.py` we set TrainingArguments.per_device_batch_size based on the number of processes. However, Trainer automatically uses all devices it can detect, so if you run training on a 2-GPU devbox even without accelerate, it will use two devices and 1 process, doing straightforward data parallelism (not FSDP). Then the total batch size summed across all devices is twice what we expect.

It's probably a mistake whenever we launch something in a devbox like this since it is not reflective of our real batch jobs, where we always have 1 GPU per process. This PR adds an assert guarding against this mistake