circleci: Remove multi-GPU flake handling (#1048)
## Describe your changes

Reverts PR #828, which added flake handling. But we resolved the particular flake that #828 was handling and changed the `FLAKE_ERROR_MESSAGES` variable to be empty, which, due to how we check for flakes with `grep`, made every multi-GPU test count as a flake that should be retried. This meant that we unnecessarily ran failing multi-GPU test runs 3 times.

This PR removes the flake handling logic. (It's possible that because we were retrying multi-GPU tests 3 times that there are other flakes that were covered up by always re-running — if so we can revert this PR and add the flakes to `FLAKE_ERROR_MESSAGES`)