experiment_utils: Fix extra (non-existent) pythia eval round (#1179)
Here are the pythia rounds output by `get_all_n_rounds_to_evaluate_pythia`:
```python
[
    [1, 2, 3, 4, 6, 9, 15, 24, 38, 60],
    [1, 2, 3, 4, 6, 9, 15, 24, 38, 60],
    [1, 2, 3, 4, 6, 9, 15, 24, 38, 60],
    [1, 2, 3, 4, 6, 9, 15, 23, 37, 59],
    [1, 2, 3, 4, 5, 6, 7, 10, 14, 21],
    [1, 2, 3, 4, 5, 6, 7, 8, 9],
    [1, 2, 3, 4, 5, 6, 7],
    [1, 2, 3, 4, 5, 6],
    [1, 2, 3, 4, 5, 6],
    [1, 2, 3, 4, 5, 6],
]
```

This has an off-by-one for the last 5 models — the last round does not exist (e.g., for 12B there is no adv round 6).

This PR fixes the off-by-one. It also gets rid of `start_rounds` and `end_rounds` because no one was using them I think and it made the code pretty hard for me to eyeball for correctness. We did include `start_rounds` and `end_rounds` in my original ICLR-deadline evals (`t005b`) but at the time it wasn't calling this function anyway.